requirements.txt
streamlit
spotipy
lyricsgenius
sentence-transformers
umap-learn
plotly
faiss-cpu
annoy
scikit-learn
pandas
numpy
matplotlib
seaborn
tqdm
requests